# openapi/stax-permissions-sets-20210321.json
{
    "components": {
        "parameters": {
...
            "FilterCreatedBy": {
                "description": "Filter the returned results using the \"CreatedBy\" attribute",
                "example": [
                    "12345678-1234-1234-1234-123456789012",
                    "87654321-4321-4321-4321-210987654321"
                ],
                "in": "query",
                "name": "created_by",
                "schema": {
                    "items": {
                        "$ref": "#/components/schemas/ID"
                    },
                    "type": "array"
                }
            },
...
            "FilterStatus": {
                "description": "Filter the returned results by \"Status\"",
                "example": [
                    "ACTIVE"
                ],
                "in": "query",
                "name": "status",
                "schema": {
                    "items": {
                        "$ref": "#/components/schemas/PermissionSetStatus"
                    },
                    "type": "array"
                }
            },
            "Limit": {
                "description": "Pagination. The maxmimum number of records to return",
                "in": "query",
                "name": "limit",
                "schema": {
                    "default": 100,
                    "format": "int64",
                    "maximum": 100,
                    "minimum": 1,
                    "type": "integer"
                }
            },
            "PageToken": {
                "description": "Encoded pagination token used to fetch the page of data.",
                "example": "ThlNjc5MjUwNDMzMAThlNjc5MjUwNDMzMA",
                "in": "query",
                "name": "page_token",
                "schema": {
                    "minLength": 1,
                    "type": "string"
                }
            },
...
        },
        "schemas": {
...
            "PermissionSetStatus": {
                "enum": [
                    "ACTIVE",
                    "DELETED"
                ],
                "type": "string"
            },
...
        },
        "securitySchemes": {
            "sigv4": {
                "in": "header",
                "name": "Authorization",
                "type": "apiKey",
                "x-amazon-apigateway-authtype": "awsSigv4"
            }
        }
    },
    "info": {
        "contact": {
            "url": "https://stax.io"
        },
        "description": "Implementation of Stax Permission Sets API",
        "termsOfService": "https://www.stax.io/legal/terms/",
        "title": "Stax Permission Sets API",
        "version": "20210321"
    },
    "openapi": "3.0.2",
    "paths": {
        "/permission-sets": {
            "get": {
                "description": "Return list of permission sets",
                "operationId": "ListPermissionSets",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Limit"
                    },
                    {
                        "$ref": "#/components/parameters/PageToken"
                    },
                    {
                        "$ref": "#/components/parameters/FilterStatus"
                    },
                    {
                        "$ref": "#/components/parameters/FilterCreatedBy"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListPermissionSets"
                                }
                            }
                        },
                        "description": "Success."
                    },
                    "429": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Too Many Requests. When too many requests are made to the resource."
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error. Error retrieving resources."
                    }
                },
                "security": [
                    {
                        "sigv4": []
                    }
                ],
                "summary": "List permission sets for an organisation",
                "tags": [
                    "PermissionSets"
                ],
                "x-amazon-apigateway-integration": {
                    "httpMethod": "POST",
                    "passthroughBehavior": "never",
                    "type": "aws_proxy",
                    "uri": {
                        "Fn::Sub": "arn:${AWS::Partition}:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${PublicAPILambda.Arn}:live/invocations"
                    }
                }
            },
            "post": {
                "description": "Create permission set",
                "operationId": "CreatePermissionSet",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreatePermissionSetRecord"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PermissionSetRecord"
                                }
                            }
                        },
                        "description": "Success."
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request. Error creating resource. Check JSON request."
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Conflict. Error creating permission set, the specified Name already exists."
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error. Error creating resource."
                    }
                },
                "security": [
                    {
                        "sigv4": []
                    }
                ],
                "summary": "Create permission set",
                "tags": [
                    "PermissionSets"
                ],
                "x-amazon-apigateway-integration": {
                    "httpMethod": "POST",
                    "passthroughBehavior": "never",
                    "type": "aws_proxy",
                    "uri": {
                        "Fn::Sub": "arn:${AWS::Partition}:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${PublicAPILambda.Arn}:live/invocations"
                    }
                }
            }
        },
        "/permission-sets/aws-managed-policies": {
            "get": {
                "description": "Returns a list of AWS managed policies.",
                "operationId": "ListAWSManagedPolicies",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListAWSManagedPolicies"
                                }
                            }
                        },
                        "description": "Success."
                    },
                    "429": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Too Many Requests. When too many requests are made to the resource."
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error. Error retrieving resources."
                    }
                },
                "security": [
                    {
                        "sigv4": []
                    }
                ],
                "summary": "List AWS managed policies",
                "tags": [
                    "PermissionSets"
                ],
                "x-amazon-apigateway-integration": {
                    "httpMethod": "POST",
                    "passthroughBehavior": "never",
                    "type": "aws_proxy",
                    "uri": {
                        "Fn::Sub": "arn:${AWS::Partition}:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${PublicAPILambda.Arn}:live/invocations"
                    }
                }
            }
        },
        "/permission-sets/roles": {
            "get": {
                "description": "Returns list of role assignments for an organisation. Optionally filter by a single account.",
                "operationId": "ListRoles",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/PageToken"
                    },
                    {
                        "$ref": "#/components/parameters/AccountID"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListRoles"
                                }
                            }
                        },
                        "description": "Success."
                    },
                    "429": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Too Many Requests. When too many requests are made to the resource."
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error. Error retrieving resources."
                    }
                },
                "security": [
                    {
                        "sigv4": []
                    }
                ],
                "summary": "List roles",
                "tags": [
                    "Roles"
                ],
                "x-amazon-apigateway-integration": {
                    "httpMethod": "POST",
                    "passthroughBehavior": "never",
                    "type": "aws_proxy",
                    "uri": {
                        "Fn::Sub": "arn:${AWS::Partition}:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${PublicAPILambda.Arn}:live/invocations"
                    }
                }
            }
        },
        "/permission-sets/roles/me": {
            "get": {
                "description": "Returns list of role assignments for the user making the request. Only applies to Users (including federated users) and does not apply to Stax API Tokens. Optionally filter by a single account.",
                "operationId": "ListMyRoles",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/PageToken"
                    },
                    {
                        "$ref": "#/components/parameters/Limit"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListRoles"
                                }
                            }
                        },
                        "description": "Success."
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Forbidden. When the requesting identity does not have access to the endpoint, for example when callled by a Stax API token."
                    },
                    "429": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Too Many Requests. When too many requests are made to the resource."
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error. Error retrieving resources."
                    }
                },
                "security": [
                    {
                        "sigv4": []
                    }
                ],
                "summary": "List my roles",
                "tags": [
                    "Roles"
                ],
                "x-amazon-apigateway-integration": {
                    "httpMethod": "POST",
                    "passthroughBehavior": "never",
                    "type": "aws_proxy",
                    "uri": {
                        "Fn::Sub": "arn:${AWS::Partition}:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${PublicAPILambda.Arn}:live/invocations"
                    }
                }
            }
        },
...
        "/public/api-document": {
            "get": {
                "description": "Get the OpenAPI specification document for this API.",
                "operationId": "GetAPIDocument",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": true,
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Success"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error. Error retrieving resources."
                    }
                },
                "summary": "Get the OpenAPI specification document for this API.",
                "tags": [
                    "Public"
                ],
                "x-amazon-apigateway-auth": {
                    "type": "NONE"
                },
                "x-amazon-apigateway-integration": {
                    "httpMethod": "POST",
                    "passthroughBehavior": "never",
                    "type": "aws_proxy",
                    "uri": {
                        "Fn::Sub": "arn:${AWS::Partition}:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${PublicAPILambda.Arn}:live/invocations"
                    }
                }
            }
        }
    },
    "servers": [
        {
            "url": "https://api.idam.au1.staxapp.cloud/20210321"
        }
    ]
}